implement Peano numbers and show in different bases
this is an exercise in constructive minimalism.
the goal was to implement natural numbers with addition and represent them as character strings in different bases, using only built-in language features.

inded, the only non-core dependency strictly required for this is `GHC.Show` to display the results.
`GHC.Types.Char` is only needed to make all type annotations possible, without which the code would be much harder to read.